# Invoke indexer in RepositoryUpdater grouped by mimepath of file