@matthiasblaesing 2022-07-02
@dbalek could you please have a look at this? You introduced the grouped mode, so you might be in the best position to check.